==> emotion_STS/melo_rlhf_realgoal/05_inference_mix.py <==
#!/home/ubuntu/OpenVoice/.openvoiceenv/bin/python
# import matplotlib.pyplot as plt
# import IPython.display as ipd
import torch
import utils
from api import STS
import os
from glob import glob
import pandas as pd


hps = utils.get_hparams_from_file("/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/logs_emo/example/config.json")
print('hps',hps)
device = "cuda:0" if torch.cuda.is_available() else "cpu"

config_path='/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/logs_emo/example/config.json'
# ckpt_path='/home/ubuntu/OpenVoice/emotion_STS/melo/logs/example/G_6600.pth'
# ckpt_path='/home/ubuntu/OpenVoice/emotion_STS/melo_emo/ckpt/G_46800.pth'
# ckpt_path='/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/logs_emo/example/G_58630.pth'
ckpt_path='/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/logs_emo/example/G_61330.pth'
model = STS(device=device,config_path=config_path,ckpt_path=ckpt_path)
speaker_ids = model.hps.data.spk2id
cwd=os.getcwd()
output_directory=ckpt_path.split('/')[-3] + '/' + ckpt_path.split('/')[-2]
output_dir=f'{cwd}/{output_directory}/evaluation/G_61330_ch'
ckpt_num=ckpt_path.split('/')[-1].split('.')[0]
if not os.path.exists(output_dir):
    os.makedirs(output_dir)

# for speaker_key in speaker_ids.keys():
# speaker_id = speaker_ids['4362']
# output_path=f'{output_dir}/real/G_58270_ch3.wav'
# output_path=f'{output_dir}/real/cremad_neutral_angry_G_58650_1_realemo.wav'
# print(output_path)
# directory_path = os.path.dirname(output_path)
# if not os.path.exists(directory_path):
#     os.makedirs(directory_path)
# if not os.path.exists(output_path):
#     data, samplerate = sf.read(file_path)

# output_path=f'{output_dir}/0013_000017_surprise_46800_new.wav'
# output_path=f'{output_dir}/32_30510_friendly_tmp25500_new.wav'
# output_path=f'{output_dir}/Angry-0013_000017_20_tmp46400_new.wav'
real_audio_directory='/home/ubuntu/OpenVoice/data/esd/ch/input'
emo_directatry='/home/ubuntu/OpenVoice/data/esd/ch/ref'
wav_pattern='**/*.wav'
emo_pattern='**/*.wav.emo.npy'
df=pd.DataFrame()
inf={}

# df=pd.DataFrame()
for file in glob(os.path.join(real_audio_directory, wav_pattern), recursive=True):
    input_name = file.split('/')[-1].split('.')[0]

    audiopath=[file]
    file_emo=file.split('/')[-1].split('_')[-1].split('.')[0].lower()
    speaker=file.split('/')[-1].split('_')[0]
    for file2 in glob(os.path.join(emo_directatry, emo_pattern), recursive=True):
        file2_ref = file2.split('/')[-1].split('.')[0]
        file2_emo=file2.split('/')[-1].split('_')[-1].split('.')[0].lower()
        # print(file2)
        if file_emo !=file2_emo:
            inf['reference_audio']=file2
            inf['input_audio']=file
            output_path=f'{output_dir}/{input_name}_to_{file2_ref}_{ckpt_num}.wav'
            inf['output_audio']=output_path
            emo=file2
            model.sts_to_file(audiopath,emo,output_path)
            df = pd.concat([df, pd.DataFrame([inf])], ignore_index=True)
df.to_csv(f'{output_dir}/eval.csv')








# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/IWW_friendly_female.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/458_2922_cheerful.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/302_61159_angry.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/2289_20782_excited.wav']


# audiopath=['/home/ubuntu/OpenVoice/data/full_dataset/4362_48858_cheerful.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf/eval_input_audio/real/0011_000232_neutral.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf/eval_input_audio/real/0011_000852_happy.wav']
# audiopath=['/home/ubuntu/efs/sts/Emotion_Speech_Dataset/0007/Angry/0007_000425.wav']
# audiopath=['/home/ubuntu/efs/sts/emo_v/bea/neutral_225-252_0251.wav']

# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/eval_input_audio/real/0011_000488_angry.wav']
# audiopath=['/home/ubuntu/efs/sts/CREMA-D/AudioWAV/1091_IOM_NEU_XX.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf/eval_input_audio/real/0014_000892_happy.wav']
# audiopath=['/home/ubuntu/OpenVoice/data/full_dataset/1054_MTI_SAD_XX.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/eval_input_audio/real/0011_001486_surprise.wav']



# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/generate_audio/4362_48858_cheerful.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/7367_30510_angry.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/5561_67864_angry.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/generate_audio/34281_angry_tts.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/generate_audio/32_30510_angry.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo_emo/generate_audio/4362_777_angry.wav']
# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/esd/input/0013_000017.wav']

# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/4362_48858_cheerful.wav']

# audiopath=['/home/ubuntu/OpenVoice/emotion_STS/esd/input/0013_000017.wav']

# emo='/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/eval_ref_audio/real/Neutral-Angry-0013_000002.wav.emo.npy'
# emo='/home/ubuntu/efs/sts/Emotion_Speech_Dataset/0007/Happy/0007_000775.wav.emo.npy'
# emo='/home/ubuntu/efs/sts/emo_v/bea/anger_169-196_0161.wav.emo.npy'

# emo='/home/ubuntu/OpenVoice/emotion_STS/melo_rlhf_realgoal/eval_ref_audio/real/Neutral-Happy-0013_000007.wav.emo.npy'
# emo='/home/ubuntu/efs/sts/CREMA-D/AudioWAV/1080_TIE_ANG_XX.wav.emo.npy'
# emo='/home/ubuntu/efs/sts/train/tts/51028_sad_tts.wav.emo.npy'
# emo='/home/ubuntu/efs/CREMA-D/AudioWAV/1001_MTI_ANG_XX.wav.emo.npy'
# emo='/home/ubuntu/OpenVoice/data/full_dataset/34281_angry_tts.wav.emo.npy'
# emo='/home/ubuntu/OpenVoice/emotion_STS/esd/emo/Neutral-Angry-0013_000020.wav.emo.npy'
# emo='/home/ubuntu/efs/sts/train/tts/34503_friendly_tts.wav.emo.npy'
# emo='/home/ubuntu/OpenVoice/emotion_STS/esd/emo/Neutral-Surprise-0013_000017.wav.emo.npy'
# emo='/home/ubuntu/OpenVoice/emotion_STS/esd/emo/Neutral-Angry-0013_000017.wav.emo.npy'

# emo='/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/42702_sad_tts.wav.emo.npy'
# emo='/home/ubuntu/OpenVoice/emotion_STS/melo/generate_audio/24542_shouting_tts.wav.emo.npy'


# model.sts_to_file(audiopath,emo,output_path)
